+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 01. :gift: Added exception `dev.orne.http.client.HttpAuthenticationFailedException`
 01. :gift: Added class `dev.orne.http.client.AbstractHttpServiceOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusIndependentOperation`
+01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentOperation`